Gestion des rôles et permissions

+++ src/Controller/ReponseController.php
@@ -3,9 +3,11 @@
 namespace App\Controller;
 
 use App\Entity\Reponse;
+use App\Entity\User;
 use App\Form\AnswerFormType;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -44,17 +46,55 @@ class ReponseController extends AbstractController
         ]);
     }
 
-    #[IsGranted('REPONSE_EDIT', 'reponse',"YOU SHALL NOT PASS!!! modification impossible")]
+    #[IsGranted('REPONSE_DELETE', 'reponse',"YOU SHALL NOT PASS!!! modification impossible")]
     #[Route('/reponse/{id}/delete', name:'delete_reponse')]
-    public function deleteReponse(Reponse $reponse, EntityManagerInterface $entityManager)
+    public function deleteReponse(Reponse $reponse, EntityManagerInterface $entityManager, Request $request)
     {  
-        $entityManager->remove($reponse); 
-        $entityManager->flush(); 
+        $token = $request->request->get('_token');
+        $method = $request->request->get('_method');
 
-        $this->addFlash('success', "Votre réponse a correctement été supprimé!");
+        if ($method === 'DELETE' && $this->isCsrfTokenValid('reponse_delete-'. $reponse->getId(), $token)){
+
+            //Effectue la suppression
+            $entityManager->remove($reponse);
+            $entityManager->flush();
+            
+            $this->addFlash('success', 'Reponse supprimée!!! ');
+
+        }else{
+            $this->addFlash('error', 'Vous ne pouvez pas supprimer cette réponse');
+        }
 
         return $this->redirectToRoute('question_details', [
             'id' => $reponse->getQuestion()->getId()
         ]);
     }
+    #[IsGranted('REPONSE_VOTE', 'reponse', 'Vous avez déjà voté')]
+    #[Route('/reponse/{id}/vote', name:'vote_reponse')]
+    public function vote(Reponse $reponse, EntityManagerInterface $entityManager, Request $request):RedirectResponse
+    {
+        $token = $request->request->get('_token');
+        
+        if($request->getMethod() === 'POST' && $this->isCsrfTokenValid('vote-'.$reponse->getId(), $token))
+        {
+            /** @var User $user */
+            $user= $this->getUser();
+
+            // Associe la réponse à l'utilisateur
+            $user->addVote($reponse);
+
+            $entityManager->persist($user);
+            $entityManager->flush();
+    
+            $this->addFlash('success','Merci pour le vote!');
+        }else{
+            $this->addFlash('error','Vous ne pouvez plus voter ici');
+        }
+        
+        return $this->redirectToRoute('question_details',[
+            'id' => $reponse->getQuestion()->getId()
+        ]);
+
+    }
+
 }
